User edit/create form improvements. - Renamed and split files, similar to how facility form files are organised. - Replaced the SimpleSelect with Select in members table.
