+++ resources/js/components/ui/form/Select.module.scss
@@ -104,7 +104,7 @@
 .selectPortal {
   > [data-part="content"] {
     @apply bg-white border rounded shadow-xl;
-    @apply overflow-x-clip overflow-y-auto;
+    @apply overflow-x-clip overflow-y-auto z-dropdown;
     max-height: var(--available-height);
     min-width: var(--reference-width);
     max-width: var(--available-width);
