+++ resources/js/features/facility-edit/FacilityForm.tsx
@@ -15,15 +15,13 @@ const getSchema = () =>
 export type FacilityFormInput = z.input<ReturnType<typeof getSchema>>;
 export type FacilityFormOutput = z.output<ReturnType<typeof getSchema>>;
 
-type FormProps = FormConfigWithoutTransformFn<FacilityFormInput>;
-type MyProps = {
-  onCancel?: () => void;
-  id: string;
-};
-type Props = FormProps & MyProps;
+interface Props extends FormConfigWithoutTransformFn<FacilityFormInput> {
+  readonly id: string;
+  readonly onCancel?: () => void;
+}
 
 export const FacilityForm: VoidComponent<Props> = (allProps) => {
-  const [props, formProps]: [MyProps, FormProps] = splitProps(allProps, ["id", "onCancel"]);
+  const [props, formProps] = splitProps(allProps, ["id", "onCancel"]);
   return (
     <FelteForm id={props.id} schema={getSchema()} {...formProps} class="flex flex-col gap-4">
       <div class="flex flex-col gap-1">
